test: split usage repository tests out of test_db_postgres.py
Mirror the qfa.adapters.db -> qfa.adapters.usage_repository source split
on the test side. Repository-level integration tests (record_call,
get_usage_stats, get_all_usage_stats and their downstream behaviours)
move to a new tests/integration/test_usage_repository.py; schema-level
tests that exercise the llm_calls table directly via pg_engine (CHECK
constraints, declared index presence) stay in test_db_postgres.py.

No test logic changes; same 15 integration tests collected before and
after the move.
